fixup! refactor(compiler): add support for host directives
