+++ packages/compiler-cli/src/ngtsc/annotations/directive/test/directive_spec.ts
@@ -11,7 +11,7 @@ import ts from 'typescript';
 import {absoluteFrom} from '../../../file_system';
 import {runInEachFileSystem} from '../../../file_system/testing';
 import {ReferenceEmitter} from '../../../imports';
-import {CompoundMetadataReader, DtsMetadataReader, HostDirectivesResolver, InjectableClassRegistry, LocalMetadataRegistry} from '../../../metadata';
+import {CompoundMetadataReader, DtsMetadataReader, InjectableClassRegistry, LocalMetadataRegistry} from '../../../metadata';
 import {PartialEvaluator} from '../../../partial_evaluator';
 import {NOOP_PERF_RECORDER} from '../../../perf';
 import {ClassDeclaration, isNamedClassDeclaration, TypeScriptReflectionHost} from '../../../reflection';
@@ -169,15 +169,13 @@ runInEachFileSystem(() => {
         metaReader, new CompoundMetadataReader([metaReader, dtsReader]),
         new MetadataDtsModuleScopeResolver(dtsReader, null), refEmitter, null);
     const injectableRegistry = new InjectableClassRegistry(reflectionHost);
-    const hostDirectivesResolver = new HostDirectivesResolver(metaReader);
     const handler = new DirectiveDecoratorHandler(
         reflectionHost, evaluator, scopeRegistry, scopeRegistry, metaReader, injectableRegistry,
         refEmitter,
         /*isCore*/ false,
         /*semanticDepGraphUpdater*/ null,
         /*annotateForClosureCompiler*/ false,
-        /*detectUndecoratedClassesWithAngularFeatures*/ false, NOOP_PERF_RECORDER,
-        hostDirectivesResolver);
+        /*detectUndecoratedClassesWithAngularFeatures*/ false, NOOP_PERF_RECORDER);
 
     const DirNode = getDeclaration(program, _('/entry.ts'), dirName, isNamedClassDeclaration);
