+++ packages/compiler-cli/src/ngtsc/metadata/BUILD.bazel
@@ -9,7 +9,6 @@ ts_library(
     ]),
     deps = [
         "//packages/compiler",
-        "//packages/compiler-cli/src/ngtsc/diagnostics",
         "//packages/compiler-cli/src/ngtsc/file_system",
         "//packages/compiler-cli/src/ngtsc/imports",
         "//packages/compiler-cli/src/ngtsc/reflection",
