+++ packages/compiler-cli/src/ngtsc/metadata/src/host_directives_resolver.ts
@@ -6,10 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import ts from 'typescript';
-
-import {ErrorCode, makeDiagnostic} from '../../diagnostics';
-import {DirectiveMeta, HostDirectiveMeta, MatchSource, MetadataReader} from '../../metadata/src/api';
+import {DirectiveMeta, MatchSource, MetadataReader} from '../../metadata/src/api';
+import {ClassDeclaration} from '../../reflection';
 import {BindingPropertyName, ClassPropertyMapping, ClassPropertyName} from '../src/property_mapping';
 
 import {flattenInheritedDirectiveMetadata} from './inheritance';
@@ -18,52 +16,21 @@ const EMPTY_ARRAY: any[] = [];
 
 /** Resolves the host directives of a directive to a flat array of matches. */
 export class HostDirectivesResolver {
+  private cache = new Map<ClassDeclaration, ReadonlyArray<DirectiveMeta>>();
+
   constructor(private metaReader: MetadataReader) {}
 
   /** Resolves all of the host directives that apply to a directive. */
   resolve(metadata: DirectiveMeta): ReadonlyArray<DirectiveMeta> {
-    return metadata.hostDirectives && metadata.hostDirectives.length > 0 ?
+    if (this.cache.has(metadata.ref.node)) {
+      return this.cache.get(metadata.ref.node)!;
+    }
+
+    const results = metadata.hostDirectives && metadata.hostDirectives.length > 0 ?
         this.walkHostDirectives(metadata.hostDirectives, []) :
         EMPTY_ARRAY;
-  }
-
-  /** Validates that the passed in array of host directives is correct. */
-  validate(hostDirectives: HostDirectiveMeta[]): ts.DiagnosticWithLocation[] {
-    if (!hostDirectives.length) {
-      return EMPTY_ARRAY;
-    }
-
-    const diagnostics: ts.DiagnosticWithLocation[] = [];
-
-    for (const current of hostDirectives) {
-      // Note that we don't need to go through `flattenInheritedDirectiveMetadata` here,
-      // because we're only validating fields that can't be inherited.
-      const hostMeta = this.metaReader.getDirectiveMetadata(current.directive);
-
-      if (hostMeta === null) {
-        diagnostics.push(makeDiagnostic(
-            ErrorCode.UNRESOLVED_HOST_DIRECTIVE,
-            current.directive.getOriginForDiagnostics(current.origin),
-            `Could not resolve host directive metadata of ${current.directive.debugName}`));
-        continue;
-      }
-
-      if (!hostMeta.isStandalone) {
-        diagnostics.push(makeDiagnostic(
-            ErrorCode.HOST_DIRECTIVE_NOT_STANDALONE,
-            current.directive.getOriginForDiagnostics(current.origin),
-            `Host directive ${hostMeta.name} must be standalone`));
-      }
-
-      if (hostMeta.isComponent) {
-        diagnostics.push(makeDiagnostic(
-            ErrorCode.HOST_DIRECTIVE_COMPONENT,
-            current.directive.getOriginForDiagnostics(current.origin),
-            `Host directive ${hostMeta.name} cannot be a component`));
-      }
-    }
-
-    return diagnostics;
+    this.cache.set(metadata.ref.node, results);
+    return results;
   }
 
   /**
@@ -76,7 +43,7 @@ export class HostDirectivesResolver {
     for (const current of directives) {
       const hostMeta = flattenInheritedDirectiveMetadata(this.metaReader, current.directive);
 
-      // This case is already handled in `validate`, but we keep the assertion here so that the
+      // This case is already have been validated for, but we keep the assertion here so that the
       // user gets a better error message than "Cannot read property foo of null" in case something
       // slipped through.
       if (hostMeta === null) {
